scraper/scraper.py
from selenium import webdriver
from bs4 import BeautifulSoup
import requests
import logging

class CoursScraper:
    def __init__(self):
        self.driver = None
        self.setup_logging()

    def setup_logging(self):
        logging.basicConfig(
            level=logging.INFO,
            format='%(asctime)s - %(levelname)s - %(message)s'
        )

    def scrape_course(self, url):
        try:
            # Logique de scraping ici
            pass
        except Exception as e:
            logging.error(f"Erreur lors du scraping: {str(e)}")