requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
pandas==2.1.4
scikit-learn==1.3.2
psycopg2-binary==2.9.9
django-allauth==0.59.0